move host and realm concepts under runtime
