+++ Jint/Native/Prototype.cs
@@ -1,4 +1,5 @@
 using Jint.Native.Object;
+using Jint.Runtime;
 
 namespace Jint.Native
 {
